refactor: User 엔티티 변경 - #3
4/6 회의를 통해 변경된 요구사항에 따라 일부 필드를 추가/삭제하고 네이밍을 변경했습니다.
변경내용:
- Users -> User
- userJwt 삭제
- userId 삭제
- userSeq -> id
- passWord -> password
- userNickName -> nickname
